requirements.txt
streamlit
opencv-python
numpy
pytesseract
python-dotenv
groq
requests
Pillow
flask